+++ test/input/beamville/beam.conf
@@ -76,7 +76,7 @@ beam.agentsim.agents.rideHail.initialLocation.name="HOME"
 beam.agentsim.agents.rideHail.initialLocation.home.radiusInMeters=10000
 
 # allocationManager(DEFAULT_MANAGER | STANFORD_V1 | BUFFERED_IMPL_TEMPLATE | RANDOM_REPOSITIONING | REPOSITIONING_LOW_WAITING_TIMES )
-beam.agentsim.agents.rideHail.allocationManager.name="RANDOM_REPOSITIONING"
+beam.agentsim.agents.rideHail.allocationManager.name="Test_beam.agentsim.agents.rideHail.allocation.examples.DummyRideHailDispatchWithBufferingRequests"
 beam.agentsim.agents.rideHail.allocationManager.timeoutInSeconds=300
 beam.agentsim.agents.rideHail.allocationManager.randomRepositioning.repositioningShare=0.2
 
@@ -104,9 +104,10 @@ beam.agentsim.agents.vehicles.bicycles.useBikes=false
 # Debugging
 ##################################################################
 beam.debug.debugEnabled = true
-beam.debug.skipOverBadActors = false
+beam.debug.skipOverBadActors = true
 beam.debug.debugActorTimerIntervalInSec = 60
 beam.debug.actor.logDepth = 12
+beam.debug.secondsToWaitForSkip=5
 ##################################################################
 # Calibration
 ##################################################################
